Create README.md
The AI read the track by reading from a more texture similar to that which is rendered, it contains starting positions and track outlines. 
The AI then use feelers to feel for the edge of the track, using the pixel colour to determine distance to the edge of the track. 


This was my submission for the Computing Project (also known as dissertaion) for my degree in Computer Games Programming. 
I recieved a first for this project and degree overall. 

I have since refactored this during the summer after graduation, and I am currently experimenting with getting neural networks to learn how to drive. 
Currently I have yet to have much success with the neural networks.